pyproject.toml
[project]
name = "04-chainlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "litellm>=1.72.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.3",
]
